# Optimize Set#&, Set#clone and Set.new