UPDATE:
I also init the size on a Union now (as the size has to be at least the Max size of the two sets, this has similar gains as changes to clone and new)

Changes to clone and new involve setting the initial capacity. Always in the case of `clone`, and an overload for `Indexable(T)` in the case of new for access to an efficient `size`.  (I also expect this to generate less garbage, so the real world gain from should be even better)

```
0.25.0
new 100 items 243.51k (  4.11µs) (± 7.58%) 
new 1 000 000 items  22.69  ( 44.08ms) (± 3.39%)
clone 100 items 239.94k (  4.17µs) (± 5.79%)
clone 1 000 000 items  21.42  ( 46.69ms) (± 3.56%)
```

```
PR
new 100 items 282.04k (  3.55µs) (± 9.10%)
new 1 000 000 items  25.95  ( 38.54ms) (± 5.07%)
clone 100 items  280.8k (  3.56µs) (± 5.76%)
clone 1 000 000 items  24.84  ( 40.26ms) (± 4.88%)
```

For `&` it's the same optimization as what `intersects?` does: iterating over the smaller of the two:

```
0.25.0
sparse & small  17.54k ( 57.01µs) (± 2.78%)
dense with small  26.63k ( 37.56µs) (± 2.99%)
small with sparse    9.4M (106.34ns) (± 3.39%)
small with dense   9.34M (107.01ns) (± 2.34%)

PR
sparse & small    7.8M ( 128.2ns) (± 2.59%)
dense with small   8.03M (124.61ns) (± 2.46%)
small with sparse   7.85M (127.32ns) (± 2.88%)
small with dense   8.07M (123.98ns) (± 2.38%)
```

small is 1 item, and both sparse and dense are in the 10K range.