@sdogruyol 2017-02-18
Am i reading this correct for `&`?

Sparse & small `17.54k` vs `7.8M` is 444 times faster 😮 ?