@karlseguin 2017-02-18
@sdogruyol 

```
small = Set.new([1])
large = Set.new(Array.new(1_000_000) {|i| i})
```

In the the existing code, if you did `small & large` you'd only iterate through `smalls` single item and check its existence in `large`. However, if you did `large & small` you'd iterate all million items and check all million of them. But in the end, both give the same result.

With the PR, it will always let the small one "lead". The cost is figuring out which is smaller (self.size < other.size)...and while intersecting 1 to 1_000_000 is extreme,  the gain should be fairly linear relative to the size difference.